refactor: harmonise file names
