+++ workflows/data_pipelines/agence_bio/README.md
@@ -4,6 +4,7 @@
 
 | Information | Valeur |
 | -------- | -------- |
+| Fichier source     | `dag.py`     |
 | Description | Ce traitement permet de constituer un stock sur les entreprises certifiées bio en France. |
 | Fréquence | Quotidienne |
 | Données sources | [API Professionnels Bio](https://api.gouv.fr/les-api/api-professionnels-bio) |
