+++ workflows/data_pipelines/agence_bio/dag.py
@@ -3,13 +3,10 @@ from datetime import timedelta
 from airflow.decorators import dag, task
 from airflow.utils.dates import days_ago
 
-from dag_datalake_sirene.config import EMAIL_LIST
 from dag_datalake_sirene.helpers import Notification
-from dag_datalake_sirene.workflows.data_pipelines.uai.uai_config import (
-    UAI_CONFIG,
-)
-from dag_datalake_sirene.workflows.data_pipelines.uai.uai_processor import (
-    UAIProcessor,
+from dag_datalake_sirene.config import EMAIL_LIST
+from dag_datalake_sirene.workflows.data_pipelines.agence_bio.processor import (
+    AgenceBioProcessor,
 )
 
 default_args = {
@@ -22,46 +19,41 @@ default_args = {
 
 
 @dag(
-    tags=["uai", "scolaire"],
+    tags=["agence bio", "certifications"],
     default_args=default_args,
     schedule_interval="0 16 * * *",
     start_date=days_ago(8),
-    dagrun_timeout=timedelta(minutes=15),
+    dagrun_timeout=timedelta(minutes=60 * 5),
     params={},
     catchup=False,
     on_failure_callback=Notification.send_notification_tchap,
     on_success_callback=Notification.send_notification_tchap,
 )
-def data_processing_uai():
-    uai_processor = UAIProcessor()
+def data_processing_agence_bio():
+    agence_bio_processor = AgenceBioProcessor()
 
     @task.bash
     def clean_previous_outputs():
-        return f"rm -rf {UAI_CONFIG.tmp_folder} && mkdir -p {UAI_CONFIG.tmp_folder}"
-
-    @task
-    def download_data():
-        return uai_processor.download_data()
+        return f"rm -rf {agence_bio_processor.config.tmp_folder} && mkdir -p {agence_bio_processor.config.tmp_folder}"
 
     @task
     def preprocess_data():
-        return uai_processor.preprocess_data()
+        return agence_bio_processor.preprocess_data()
 
     @task
     def save_date_last_modified():
-        return uai_processor.save_date_last_modified()
+        return agence_bio_processor.save_date_last_modified()
 
     @task
     def send_file_to_minio():
-        return uai_processor.send_file_to_minio()
+        return agence_bio_processor.send_file_to_minio()
 
     @task
     def compare_files_minio():
-        return uai_processor.compare_files_minio()
+        return agence_bio_processor.compare_files_minio()
 
     (
         clean_previous_outputs()
-        >> download_data()
         >> preprocess_data()
         >> save_date_last_modified()
         >> send_file_to_minio()
@@ -69,4 +61,4 @@ def data_processing_uai():
     )
 
 
-data_processing_uai()
+data_processing_agence_bio()
