+++ workflows/data_pipelines/bilans_financiers/README.md
@@ -4,7 +4,7 @@
 
 | Information | Valeur |
 | -------- | -------- |
-| Fichier source     | `bilans_financiers_dag.py`     |
+| Fichier source     | `dag.py`     |
 | Description | Ce traitement permet de synthétiser les données publiées par l'équipe Signaux Faibles et Fiche Commune Entreprise. |
 | Fréquence | Quotidienne |
 | Données sources | [Jeu de données INPI ratios financiers](https://www.data.gouv.fr/fr/datasets/63cb2e29b22886911440440d/) |
