+++ workflows/data_pipelines/colter/dag.py
@@ -6,11 +6,11 @@ from airflow.datasets import Dataset
 
 from dag_datalake_sirene.config import EMAIL_LIST
 from dag_datalake_sirene.helpers import Notification
-from dag_datalake_sirene.workflows.data_pipelines.colter.colter_config import (
+from dag_datalake_sirene.workflows.data_pipelines.colter.config import (
     COLTER_CONFIG,
     ELUS_CONFIG,
 )
-from dag_datalake_sirene.workflows.data_pipelines.colter.colter_processor import (
+from dag_datalake_sirene.workflows.data_pipelines.colter.processor import (
     ColterProcessor,
     ElusProcessor,
 )
