+++ workflows/data_pipelines/ess_france/dag.py
@@ -1,16 +1,15 @@
-from datetime import timedelta
-
 from airflow.decorators import dag, task
 from airflow.utils.dates import days_ago
-
+from datetime import timedelta
 from dag_datalake_sirene.config import EMAIL_LIST
 from dag_datalake_sirene.helpers import Notification
-from dag_datalake_sirene.workflows.data_pipelines.finess.config import FINESS_CONFIG
-from dag_datalake_sirene.workflows.data_pipelines.finess.finess_processor import (
-    FinessProcessor,
+from dag_datalake_sirene.workflows.data_pipelines.ess_france.processor import (
+    EssFranceProcessor,
 )
+from dag_datalake_sirene.workflows.data_pipelines.ess_france.config import ESS_CONFIG
 
-finess_processor = FinessProcessor()
+ess_france_processor = EssFranceProcessor()
+
 default_args = {
     "depends_on_past": False,
     "email_on_failure": True,
@@ -21,7 +20,7 @@ default_args = {
 
 
 @dag(
-    tags=["domaine sanitaire et social"],
+    tags=["economie sociale et solidaire", "ESS France"],
     default_args=default_args,
     schedule_interval="0 16 * * *",
     start_date=days_ago(8),
@@ -31,41 +30,35 @@ default_args = {
     on_failure_callback=Notification.send_notification_tchap,
     on_success_callback=Notification.send_notification_tchap,
 )
-def data_processing_finess():
+def data_processing_ess_france():
     @task.bash
     def clean_previous_outputs():
-        return (
-            f"rm -rf {FINESS_CONFIG.tmp_folder} && mkdir -p {FINESS_CONFIG.tmp_folder}"
-        )
+        return f"rm -rf {ESS_CONFIG.tmp_folder} && mkdir -p {ESS_CONFIG.tmp_folder}"
 
     @task
-    def download_data():
-        return finess_processor.download_data()
-
-    @task
-    def preprocess_finess():
-        return finess_processor.preprocess_data()
+    def preprocess_ess_france():
+        return ess_france_processor.preprocess_data()
 
     @task
     def save_date_last_modified():
-        return finess_processor.save_date_last_modified()
+        return ess_france_processor.save_date_last_modified()
 
     @task
     def send_file_to_minio():
-        return finess_processor.send_file_to_minio()
+        return ess_france_processor.send_file_to_minio()
 
     @task
     def compare_files_minio():
-        return finess_processor.compare_files_minio()
+        return ess_france_processor.compare_files_minio()
 
     (
         clean_previous_outputs()
-        >> download_data()
-        >> preprocess_finess()
+        >> preprocess_ess_france()
         >> save_date_last_modified()
         >> send_file_to_minio()
         >> compare_files_minio()
     )
 
 
-data_processing_finess()
+# Instantiate the DAG
+data_processing_ess_france()
