+++ workflows/data_pipelines/etl/data_fetch_clean/bilan_financier.py
@@ -1,5 +1,5 @@
 import pandas as pd
-from dag_datalake_sirene.workflows.data_pipelines.bilans_financiers.bilans_financiers_config import (
+from dag_datalake_sirene.workflows.data_pipelines.bilans_financiers.config import (
     BILANS_FINANCIERS_CONFIG,
 )
