+++ workflows/data_pipelines/etl/data_fetch_clean/organisme_formation.py
@@ -1,6 +1,6 @@
 import pandas as pd
 
-from dag_datalake_sirene.workflows.data_pipelines.formation.formation_config import (
+from dag_datalake_sirene.workflows.data_pipelines.formation.config import (
     FORMATION_CONFIG,
 )
