+++ workflows/data_pipelines/sirene/stock/processor.py
