+++ workflows/data_pipelines/spectacle/dag.py
@@ -4,11 +4,11 @@ from airflow.decorators import dag, task
 from airflow.utils.dates import days_ago
 
 from dag_datalake_sirene.helpers import Notification
-from dag_datalake_sirene.workflows.data_pipelines.spectacle.spectacle_config import (
+from dag_datalake_sirene.workflows.data_pipelines.spectacle.config import (
     SPECTACLE_CONFIG,
 )
 from dag_datalake_sirene.config import EMAIL_LIST
-from dag_datalake_sirene.workflows.data_pipelines.spectacle.spectacle_processor import (
+from dag_datalake_sirene.workflows.data_pipelines.spectacle.processor import (
     SpectacleProcessor,
 )
