+++ workflows/data_pipelines/spectacle/processor.py
@@ -5,7 +5,7 @@ from dag_datalake_sirene.helpers import DataProcessor, Notification
 
 class SpectacleProcessor(DataProcessor):
     def __init__(self):
-        from dag_datalake_sirene.workflows.data_pipelines.spectacle.spectacle_config import (
+        from dag_datalake_sirene.workflows.data_pipelines.spectacle.config import (
             SPECTACLE_CONFIG,
         )
